@@ -1,19 +1,19 @@
-require File.expand_path('../active_record_spec_helper', __FILE__)
+require "spec_helper"
 
-describe Audited::Auditor, :adapter => :active_record do
+describe Audited::Auditor do
 
   describe "configuration" do
     it "should include instance methods" do
-      Models::ActiveRecord::User.new.should be_a_kind_of( Audited::Auditor::AuditedInstanceMethods)
+      expect(Models::ActiveRecord::User.new).to be_a_kind_of( Audited::Auditor::AuditedInstanceMethods)
     end
 
     it "should include class methods" do
-      Models::ActiveRecord::User.should be_a_kind_of( Audited::Auditor::AuditedClassMethods )
+      expect(Models::ActiveRecord::User).to be_a_kind_of( Audited::Auditor::AuditedClassMethods )
     end
 
     ['created_at', 'updated_at', 'created_on', 'updated_on', 'lock_version', 'id', 'password'].each do |column|
       it "should not audit #{column}" do
-        Models::ActiveRecord::User.non_audited_columns.should include(column)
+        expect(Models::ActiveRecord::User.non_audited_columns).to include(column)
       end
     end
 
@@ -23,11 +23,11 @@ describe Audited::Auditor, :adapter => :active_record do
         audited
       end
 
-      Secret.non_audited_columns.should include('delta', 'top_secret', 'created_at')
+      expect(Secret.non_audited_columns).to include('delta', 'top_secret', 'created_at')
     end
 
     it "should not save non-audited columns" do
-      create_active_record_user.audits.first.audited_changes.keys.any? { |col| ['created_at', 'updated_at', 'password'].include?( col ) }.should be_false
+      expect(create_user.audits.first.audited_changes.keys.any? { |col| ['created_at', 'updated_at', 'password'].include?( col ) }).to eq(false)
     end
   end
 
@@ -42,17 +42,17 @@ describe Audited::Auditor, :adapter => :active_record do
                                         :suspended_at => yesterday,
                                         :logins       => 2)
 
-      u.name.should eq('name')
-      u.username.should eq('username')
-      u.password.should eq('password')
-      u.activated.should eq(true)
-      u.suspended_at.should eq(yesterday)
-      u.logins.should eq(2)
+      expect(u.name).to eq('name')
+      expect(u.username).to eq('username')
+      expect(u.password).to eq('password')
+      expect(u.activated).to eq(true)
+      expect(u.suspended_at.to_i).to eq(yesterday.to_i)
+      expect(u.logins).to eq(2)
     end
   end
 
   describe "on create" do
-    let( :user ) { create_active_record_user :audit_comment => "Create" }
+    let( :user ) { create_user :audit_comment => "Create" }
 
     it "should change the audit count" do
       expect {
@@ -61,28 +61,28 @@ describe Audited::Auditor, :adapter => :active_record do
     end
 
     it "should create associated audit" do
-      user.audits.count.should be(1)
+      expect(user.audits.count).to eq(1)
     end
 
     it "should set the action to create" do
-      user.audits.first.action.should == 'create'
-      Audited.audit_class.creates.reorder(:id).last.should == user.audits.first
-      user.audits.creates.count.should == 1
-      user.audits.updates.count.should == 0
-      user.audits.destroys.count.should == 0
+      expect(user.audits.first.action).to eq('create')
+      expect(Audited.audit_class.creates.order(:id).last).to eq(user.audits.first)
+      expect(user.audits.creates.count).to eq(1)
+      expect(user.audits.updates.count).to eq(0)
+      expect(user.audits.destroys.count).to eq(0)
     end
 
     it "should store all the audited attributes" do
-      user.audits.first.audited_changes.should == user.audited_attributes
+      expect(user.audits.first.audited_changes).to eq(user.audited_attributes)
     end
 
     it "should store comment" do
-      user.audits.first.comment.should == 'Create'
+      expect(user.audits.first.comment).to eq('Create')
     end
 
     it "should not audit an attribute which is excepted if specified on create or destroy" do
       on_create_destroy_except_name = Models::ActiveRecord::OnCreateDestroyExceptName.create(:name => 'Bart')
-      on_create_destroy_except_name.audits.first.audited_changes.keys.any?{|col| ['name'].include? col}.should be_false
+      expect(on_create_destroy_except_name.audits.first.audited_changes.keys.any?{|col| ['name'].include? col}).to eq(false)
     end
 
     it "should not save an audit if only specified on update/destroy" do
@@ -94,7 +94,7 @@ describe Audited::Auditor, :adapter => :active_record do
 
   describe "on update" do
     before do
-      @user = create_active_record_user( :name => 'Brandon', :audit_comment => 'Update' )
+      @user = create_user( :name => 'Brandon', :audit_comment => 'Update' )
     end
 
     it "should save an audit" do
@@ -108,18 +108,18 @@ describe Audited::Auditor, :adapter => :active_record do
 
     it "should set the action to 'update'" do
       @user.update_attributes :name => 'Changed'
-      @user.audits.last.action.should == 'update'
-      Audited.audit_class.updates.reorder(:id).last.should == @user.audits.last
-      @user.audits.updates.last.should == @user.audits.last
+      expect(@user.audits.last.action).to eq('update')
+      expect(Audited.audit_class.updates.order(:id).last).to eq(@user.audits.last)
+      expect(@user.audits.updates.last).to eq(@user.audits.last)
     end
 
     it "should store the changed attributes" do
       @user.update_attributes :name => 'Changed'
-      @user.audits.last.audited_changes.should == { 'name' => ['Brandon', 'Changed'] }
+      expect(@user.audits.last.audited_changes).to eq({ 'name' => ['Brandon', 'Changed'] })
     end
 
     it "should store audit comment" do
-      @user.audits.last.comment.should == 'Update'
+      expect(@user.audits.last.comment).to eq('Update')
     end
 
     it "should not save an audit if only specified on create/destroy" do
@@ -154,7 +154,7 @@ describe Audited::Auditor, :adapter => :active_record do
 
   describe "on destroy" do
     before do
-      @user = create_active_record_user
+      @user = create_user
     end
 
     it "should save an audit" do
@@ -162,21 +162,21 @@ describe Audited::Auditor, :adapter => :active_record do
         @user.destroy
       }.to change( Audited.audit_class, :count )
 
-      @user.audits.size.should be(2)
+      expect(@user.audits.size).to eq(2)
     end
 
     it "should set the action to 'destroy'" do
       @user.destroy
 
-      @user.audits.last.action.should == 'destroy'
-      Audited.audit_class.destroys.reorder(:id).last.should == @user.audits.last
-      @user.audits.destroys.last.should == @user.audits.last
+      expect(@user.audits.last.action).to eq('destroy')
+      expect(Audited.audit_class.destroys.order(:id).last).to eq(@user.audits.last)
+      expect(@user.audits.destroys.last).to eq(@user.audits.last)
     end
 
     it "should store all of the audited attributes" do
       @user.destroy
 
-      @user.audits.last.audited_changes.should == @user.audited_attributes
+      expect(@user.audits.last.audited_changes).to eq(@user.audited_attributes)
     end
 
     it "should be able to reconstruct a destroyed record without history" do
@@ -184,7 +184,7 @@ describe Audited::Auditor, :adapter => :active_record do
       @user.destroy
 
       revision = @user.audits.first.revision
-      revision.name.should == @user.name
+      expect(revision.name).to eq(@user.name)
     end
 
     it "should not save an audit if only specified on create/update" do
@@ -203,7 +203,19 @@ describe Audited::Auditor, :adapter => :active_record do
         owner.destroy
       }.to change( Audited.audit_class, :count )
 
-      company.audits.map { |a| a.action }.should == ['create', 'destroy']
+      expect(company.audits.map { |a| a.action }).to eq(['create', 'destroy'])
+    end
+  end
+
+  describe "on destroy with unsaved object" do
+    let(:user) { Models::ActiveRecord::User.new }
+
+    it "should not audit on 'destroy'" do
+      expect {
+        user.destroy
+      }.to_not raise_error
+
+      expect( user.audits ).to be_empty
     end
   end
 
@@ -212,17 +224,17 @@ describe Audited::Auditor, :adapter => :active_record do
     let(:owned_company) { Models::ActiveRecord::OwnedCompany.create!(:name => 'The auditors', :owner => owner) }
 
     it "should record the associated object on create" do
-      owned_company.audits.first.associated.should == owner
+      expect(owned_company.audits.first.associated).to eq(owner)
     end
 
     it "should store the associated object on update" do
       owned_company.update_attribute(:name, 'The Auditors')
-      owned_company.audits.last.associated.should == owner
+      expect(owned_company.audits.last.associated).to eq(owner)
     end
 
     it "should store the associated object on destroy" do
       owned_company.destroy
-      owned_company.audits.last.associated.should == owner
+      expect(owned_company.audits.last.associated).to eq(owner)
     end
   end
 
@@ -231,8 +243,8 @@ describe Audited::Auditor, :adapter => :active_record do
     let!(:owned_company) { Models::ActiveRecord::OwnedCompany.create!(:name => 'The auditors', :owner => owner) }
 
     it "should list the associated audits" do
-      owner.associated_audits.length.should == 1
-      owner.associated_audits.first.auditable.should == owned_company
+      expect(owner.associated_audits.length).to eq(1)
+      expect(owner.associated_audits.first.auditable).to eq(owned_company)
     end
   end
 
@@ -240,16 +252,16 @@ describe Audited::Auditor, :adapter => :active_record do
     let( :user ) { create_versions }
 
     it "should return an Array of Users" do
-      user.revisions.should be_a_kind_of( Array )
-      user.revisions.each { |version| version.should be_a_kind_of Models::ActiveRecord::User }
+      expect(user.revisions).to be_a_kind_of( Array )
+      user.revisions.each { |version| expect(version).to be_a_kind_of Models::ActiveRecord::User }
     end
 
     it "should have one revision for a new record" do
-      create_active_record_user.revisions.size.should be(1)
+      expect(create_user.revisions.size).to eq(1)
     end
 
     it "should have one revision for each audit" do
-      user.audits.size.should eql( user.revisions.size )
+      expect(user.audits.size).to eql( user.revisions.size )
     end
 
     it "should set the attributes for each revision" do
@@ -257,16 +269,16 @@ describe Audited::Auditor, :adapter => :active_record do
       u.update_attributes :name => 'Foobar'
       u.update_attributes :name => 'Awesome', :username => 'keepers'
 
-      u.revisions.size.should == 3
+      expect(u.revisions.size).to eql(3)
 
-      u.revisions[0].name.should == 'Brandon'
-      u.revisions[0].username.should == 'brandon'
+      expect(u.revisions[0].name).to eql('Brandon')
+      expect(u.revisions[0].username).to eql('brandon')
 
-      u.revisions[1].name.should == 'Foobar'
-      u.revisions[1].username.should == 'brandon'
+      expect(u.revisions[1].name).to eql('Foobar')
+      expect(u.revisions[1].username).to eql('brandon')
 
-      u.revisions[2].name.should == 'Awesome'
-      u.revisions[2].username.should == 'keepers'
+      expect(u.revisions[2].name).to eql('Awesome')
+      expect(u.revisions[2].username).to eql('keepers')
     end
 
     it "access to only recent revisions" do
@@ -274,18 +286,18 @@ describe Audited::Auditor, :adapter => :active_record do
       u.update_attributes :name => 'Foobar'
       u.update_attributes :name => 'Awesome', :username => 'keepers'
 
-      u.revisions(2).size.should == 2
+      expect(u.revisions(2).size).to eq(2)
 
-      u.revisions(2)[0].name.should == 'Foobar'
-      u.revisions(2)[0].username.should == 'brandon'
+      expect(u.revisions(2)[0].name).to eq('Foobar')
+      expect(u.revisions(2)[0].username).to eq('brandon')
 
-      u.revisions(2)[1].name.should == 'Awesome'
-      u.revisions(2)[1].username.should == 'keepers'
+      expect(u.revisions(2)[1].name).to eq('Awesome')
+      expect(u.revisions(2)[1].username).to eq('keepers')
     end
 
     it "should be empty if no audits exist" do
       user.audits.delete_all
-      user.revisions.should be_empty
+      expect(user.revisions).to be_empty
     end
 
     it "should ignore attributes that have been deleted" do
@@ -298,27 +310,27 @@ describe Audited::Auditor, :adapter => :active_record do
     let( :user ) { create_versions(5) }
 
     it "should maintain identity" do
-      user.revision(1).should == user
+      expect(user.revision(1)).to eq(user)
     end
 
     it "should find the given revision" do
       revision = user.revision(3)
-      revision.should be_a_kind_of( Models::ActiveRecord::User )
-      revision.version.should be(3)
-      revision.name.should == 'Foobar 3'
+      expect(revision).to be_a_kind_of( Models::ActiveRecord::User )
+      expect(revision.version).to eq(3)
+      expect(revision.name).to eq('Foobar 3')
     end
 
     it "should find the previous revision with :previous" do
       revision = user.revision(:previous)
-      revision.version.should be(4)
-      #revision.should == user.revision(4)
-      revision.attributes.should == user.revision(4).attributes
+      expect(revision.version).to eq(4)
+      #expect(revision).to eq(user.revision(4))
+      expect(revision.attributes).to eq(user.revision(4).attributes)
     end
 
     it "should be able to get the previous revision repeatedly" do
       previous = user.revision(:previous)
-      previous.version.should be(4)
-      previous.revision(:previous).version.should be(3)
+      expect(previous.version).to eq(4)
+      expect(previous.revision(:previous).version).to eq(3)
     end
 
     it "should be able to set protected attributes" do
@@ -326,14 +338,14 @@ describe Audited::Auditor, :adapter => :active_record do
       u.update_attribute :logins, 1
       u.update_attribute :logins, 2
 
-      u.revision(3).logins.should be(2)
-      u.revision(2).logins.should be(1)
-      u.revision(1).logins.should be(0)
+      expect(u.revision(3).logins).to eq(2)
+      expect(u.revision(2).logins).to eq(1)
+      expect(u.revision(1).logins).to eq(0)
     end
 
     it "should set attributes directly" do
       u = Models::ActiveRecord::User.create(:name => '<Joe>')
-      u.revision(1).name.should == '&lt;Joe&gt;'
+      expect(u.revision(1).name).to eq('&lt;Joe&gt;')
     end
 
     it "should set the attributes for each revision" do
@@ -341,20 +353,20 @@ describe Audited::Auditor, :adapter => :active_record do
       u.update_attributes :name => 'Foobar'
       u.update_attributes :name => 'Awesome', :username => 'keepers'
 
-      u.revision(3).name.should == 'Awesome'
-      u.revision(3).username.should == 'keepers'
+      expect(u.revision(3).name).to eq('Awesome')
+      expect(u.revision(3).username).to eq('keepers')
 
-      u.revision(2).name.should == 'Foobar'
-      u.revision(2).username.should == 'brandon'
+      expect(u.revision(2).name).to eq('Foobar')
+      expect(u.revision(2).username).to eq('brandon')
 
-      u.revision(1).name.should == 'Brandon'
-      u.revision(1).username.should == 'brandon'
+      expect(u.revision(1).name).to eq('Brandon')
+      expect(u.revision(1).username).to eq('brandon')
     end
 
     it "should be able to get time for first revision" do
-      suspended_at = Time.now
+      suspended_at = Time.zone.now
       u = Models::ActiveRecord::User.create(:suspended_at => suspended_at)
-      u.revision(1).suspended_at.should == suspended_at
+      expect(u.revision(1).suspended_at.to_s).to eq(suspended_at.to_s)
     end
 
     it "should not raise an error when no previous audits exist" do
@@ -363,7 +375,7 @@ describe Audited::Auditor, :adapter => :active_record do
     end
 
     it "should mark revision's attributes as changed" do
-      user.revision(1).name_changed?.should be_true
+      expect(user.revision(1).name_changed?).to eq(true)
     end
 
     it "should record new audit when saving revision" do
@@ -381,18 +393,18 @@ describe Audited::Auditor, :adapter => :active_record do
   end
 
   describe "revision_at" do
-    let( :user ) { create_active_record_user }
+    let( :user ) { create_user }
 
     it "should find the latest revision before the given time" do
       audit = user.audits.first
       audit.created_at = 1.hour.ago
       audit.save!
       user.update_attributes :name => 'updated'
-      user.revision_at( 2.minutes.ago ).version.should be(1)
+      expect(user.revision_at( 2.minutes.ago ).version).to eq(1)
     end
 
     it "should be nil if given a time before audits" do
-      user.revision_at( 1.week.ago ).should be_nil
+      expect(user.revision_at( 1.week.ago )).to be_nil
     end
   end
 
@@ -400,7 +412,7 @@ describe Audited::Auditor, :adapter => :active_record do
     it "should not save an audit when calling #save_without_auditing" do
       expect {
         u = Models::ActiveRecord::User.new(:name => 'Brandon')
-        u.save_without_auditing.should be_true
+        expect(u.save_without_auditing).to eq(true)
       }.to_not change( Audited.audit_class, :count )
     end
 
@@ -409,22 +421,27 @@ describe Audited::Auditor, :adapter => :active_record do
         Models::ActiveRecord::User.without_auditing { Models::ActiveRecord::User.create!( :name => 'Brandon' ) }
       }.to_not change( Audited.audit_class, :count )
     end
+
+    it "should reset auditing status even it raises an exception" do
+      Models::ActiveRecord::User.without_auditing { raise } rescue nil
+      expect(Models::ActiveRecord::User.auditing_enabled).to eq(true)
+    end
   end
 
   describe "comment required" do
 
     describe "on create" do
       it "should not validate when audit_comment is not supplied" do
-        Models::ActiveRecord::CommentRequiredUser.new.should_not be_valid
+        expect(Models::ActiveRecord::CommentRequiredUser.new).not_to be_valid
       end
 
       it "should validate when audit_comment is supplied" do
-        Models::ActiveRecord::CommentRequiredUser.new( :audit_comment => 'Create').should be_valid
+        expect(Models::ActiveRecord::CommentRequiredUser.new( :audit_comment => 'Create')).to be_valid
       end
 
       it "should validate when audit_comment is not supplied, and auditing is disabled" do
         Models::ActiveRecord::CommentRequiredUser.disable_auditing
-        Models::ActiveRecord::CommentRequiredUser.new.should be_valid
+        expect(Models::ActiveRecord::CommentRequiredUser.new).to be_valid
         Models::ActiveRecord::CommentRequiredUser.enable_auditing
       end
     end
@@ -433,16 +450,16 @@ describe Audited::Auditor, :adapter => :active_record do
       let( :user ) { Models::ActiveRecord::CommentRequiredUser.create!( :audit_comment => 'Create' ) }
 
       it "should not validate when audit_comment is not supplied" do
-        user.update_attributes(:name => 'Test').should be_false
+        expect(user.update_attributes(:name => 'Test')).to eq(false)
       end
 
       it "should validate when audit_comment is supplied" do
-        user.update_attributes(:name => 'Test', :audit_comment => 'Update').should be_true
+        expect(user.update_attributes(:name => 'Test', :audit_comment => 'Update')).to eq(true)
       end
 
       it "should validate when audit_comment is not supplied, and auditing is disabled" do
         Models::ActiveRecord::CommentRequiredUser.disable_auditing
-        user.update_attributes(:name => 'Test').should be_true
+        expect(user.update_attributes(:name => 'Test')).to eq(true)
         Models::ActiveRecord::CommentRequiredUser.enable_auditing
       end
     end
@@ -451,17 +468,17 @@ describe Audited::Auditor, :adapter => :active_record do
       let( :user ) { Models::ActiveRecord::CommentRequiredUser.create!( :audit_comment => 'Create' )}
 
       it "should not validate when audit_comment is not supplied" do
-        user.destroy.should be_false
+        expect(user.destroy).to eq(false)
       end
 
       it "should validate when audit_comment is supplied" do
         user.audit_comment = "Destroy"
-        user.destroy.should == user
+        expect(user.destroy).to eq(user)
       end
 
       it "should validate when audit_comment is not supplied, and auditing is disabled" do
         Models::ActiveRecord::CommentRequiredUser.disable_auditing
-        user.destroy.should == user
+        expect(user.destroy).to eq(user)
         Models::ActiveRecord::CommentRequiredUser.enable_auditing
       end
     end
@@ -492,7 +509,7 @@ describe Audited::Auditor, :adapter => :active_record do
         company.update_attributes :name => 'The Auditors'
 
         company.audits.each do |audit|
-          audit.user.should == user
+          expect(audit.user).to eq(user)
         end
       end
     end
@@ -503,7 +520,7 @@ describe Audited::Auditor, :adapter => :active_record do
         company.update_attributes :name => 'The Auditors'
 
         company.audits.each do |audit|
-          audit.user.should == user.name
+          expect(audit.user).to eq(user.name)
         end
       end
     end
@@ -513,10 +530,19 @@ describe Audited::Auditor, :adapter => :active_record do
     let( :user ) { user = Models::ActiveRecord::UserWithAfterAudit.new }
 
     it "should invoke after_audit callback on create" do
-      user.bogus_attr.should == nil
-      user.save.should == true
-      user.bogus_attr.should == "do something"
+      expect(user.bogus_attr).to be_nil
+      expect(user.save).to eq(true)
+      expect(user.bogus_attr).to eq("do something")
     end
   end
 
+  describe "around_audit" do
+    let( :user ) { user = Models::ActiveRecord::UserWithAfterAudit.new }
+
+    it "should invoke around_audit callback on create" do
+      expect(user.around_attr).to be_nil
+      expect(user.save).to eq(true)
+      expect(user.around_attr).to eq(user.audits.last)
+    end
+  end
 end
