@@ -1,14 +1,15 @@
-require File.expand_path('../active_record_spec_helper', __FILE__)
+require "spec_helper"
 
 class AuditsController < ActionController::Base
   def audit
     @company = Models::ActiveRecord::Company.create
-    render :nothing => true
+    head :ok
   end
+  attr_reader :company
 
   def update_user
-    current_user.update_attributes( :password => 'foo')
-    render :nothing => true
+    current_user.update_attributes(password: 'foo')
+    head :ok
   end
 
   private
@@ -17,8 +18,9 @@ class AuditsController < ActionController::Base
   attr_accessor :custom_user
 end
 
-describe AuditsController, :adapter => :active_record do
+describe AuditsController do
   include RSpec::Rails::ControllerExampleGroup
+  render_views
 
   before(:each) do
     Audited.current_user_method = :current_user
@@ -30,12 +32,11 @@ describe AuditsController, :adapter => :active_record do
 
     it "should audit user" do
       controller.send(:current_user=, user)
-
       expect {
         post :audit
       }.to change( Audited.audit_class, :count )
 
-      assigns(:company).audits.last.user.should == user
+      expect(controller.company.audits.last.user).to eq(user)
     end
 
     it "should support custom users for sweepers" do
@@ -46,7 +47,7 @@ describe AuditsController, :adapter => :active_record do
         post :audit
       }.to change( Audited.audit_class, :count )
 
-      assigns(:company).audits.last.user.should == user
+      expect(controller.company.audits.last.user).to eq(user)
     end
 
     it "should record the remote address responsible for the change" do
@@ -55,7 +56,16 @@ describe AuditsController, :adapter => :active_record do
 
       post :audit
 
-      assigns(:company).audits.last.remote_address.should == '1.2.3.4'
+      expect(controller.company.audits.last.remote_address).to eq('1.2.3.4')
+    end
+
+    it "should record a UUID for the web request responsible for the change" do
+      allow_any_instance_of(ActionDispatch::Request).to receive(:uuid).and_return("abc123")
+      controller.send(:current_user=, user)
+
+      post :audit
+
+      expect(controller.company.audits.last.request_uuid).to eq("abc123")
     end
 
   end
@@ -74,24 +84,24 @@ describe AuditsController, :adapter => :active_record do
 end
 
 
-describe Audited::Sweeper, :adapter => :active_record do
+describe Audited::Sweeper do
 
   it "should be thread-safe" do
     t1 = Thread.new do
       sleep 0.5
       Audited::Sweeper.instance.controller = 'thread1 controller instance'
-      Audited::Sweeper.instance.controller.should eq('thread1 controller instance')
+      expect(Audited::Sweeper.instance.controller).to eq('thread1 controller instance')
     end
 
     t2 = Thread.new do
       Audited::Sweeper.instance.controller = 'thread2 controller instance'
       sleep 1
-      Audited::Sweeper.instance.controller.should eq('thread2 controller instance')
+      expect(Audited::Sweeper.instance.controller).to eq('thread2 controller instance')
     end
 
     t1.join; t2.join
 
-    Audited::Sweeper.instance.controller.should be_nil
+    expect(Audited::Sweeper.instance.controller).to be_nil
   end
 
 end
