@@ -4,9 +4,9 @@ require File.expand_path('../schema', __FILE__)
 module Models
   module ActiveRecord
     class User < ::ActiveRecord::Base
-      audited :allow_mass_assignment => true, :except => :password
+      audited allow_mass_assignment: true, except: :password
 
-      attr_protected :logins
+      attr_protected :logins if respond_to?(:attr_protected)
 
       def name=(val)
         write_attribute(:name, CGI.escapeHTML(val))
@@ -15,33 +15,37 @@ module Models
 
     class CommentRequiredUser < ::ActiveRecord::Base
       self.table_name = :users
-      audited :comment_required => true
+      audited comment_required: true
     end
 
     class AccessibleAfterDeclarationUser < ::ActiveRecord::Base
       self.table_name = :users
       audited
-      attr_accessible :name, :username, :password
+      attr_accessible :name, :username, :password if respond_to?(:attr_accessible)
     end
 
     class AccessibleBeforeDeclarationUser < ::ActiveRecord::Base
       self.table_name = :users
-      attr_accessible :name, :username, :password # declare attr_accessible before calling aaa
+      attr_accessible :name, :username, :password if respond_to?(:attr_accessible) # declare attr_accessible before calling aaa
       audited
     end
 
     class NoAttributeProtectionUser < ::ActiveRecord::Base
       self.table_name = :users
-      audited :allow_mass_assignment => true
+      audited allow_mass_assignment: true
     end
 
     class UserWithAfterAudit < ::ActiveRecord::Base
       self.table_name = :users
       audited
-      attr_accessor :bogus_attr
+      attr_accessor :bogus_attr, :around_attr
 
       def after_audit
         self.bogus_attr = "do something"
+      end
+
+      def around_audit
+        self.around_attr = yield
       end
     end
 
@@ -57,29 +61,29 @@ module Models
 
     class OwnedCompany < ::ActiveRecord::Base
       self.table_name = 'companies'
-      belongs_to :owner, :class_name => "Owner"
-      attr_accessible :name, :owner # declare attr_accessible before calling aaa
-      audited :associated_with => :owner
+      belongs_to :owner, class_name: "Owner"
+      attr_accessible :name, :owner if respond_to?(:attr_accessible) # declare attr_accessible before calling aaa
+      audited associated_with: :owner
     end
 
     class OnUpdateDestroy < ::ActiveRecord::Base
       self.table_name = 'companies'
-      audited :on => [:update, :destroy]
+      audited on: [:update, :destroy]
     end
 
     class OnCreateDestroy < ::ActiveRecord::Base
       self.table_name = 'companies'
-      audited :on => [:create, :destroy]
+      audited on: [:create, :destroy]
     end
 
     class OnCreateDestroyExceptName < ::ActiveRecord::Base
       self.table_name = 'companies'
-      audited :except => :name, :on => [:create, :destroy]
+      audited except: :name, on: [:create, :destroy]
     end
 
     class OnCreateUpdate < ::ActiveRecord::Base
       self.table_name = 'companies'
-      audited :on => [:create, :update]
+      audited on: [:create, :update]
     end
   end
 end
