@@ -1,6 +1,7 @@
-module Audited
-  VERSION = '3.0.0'
+require 'rails/observers/active_model/active_model'
+require 'active_record'
 
+module Audited
   class << self
     attr_accessor :ignored_attributes, :current_user_method, :audit_class
 
@@ -13,3 +14,12 @@ module Audited
 
   @current_user_method = :current_user
 end
+
+require 'audited/auditor'
+require 'audited/audit'
+
+::ActiveRecord::Base.send :include, Audited::Auditor
+
+Audited.audit_class = Audited::Audit
+
+require 'audited/sweeper'
