@@ -76,6 +76,8 @@ module Audited
         "Did not expect #{@expectation}"
       end
 
+      alias_method :failure_message_when_negated, :negative_failure_message
+
       def description
         description = "audited"
         description += " associated with #{@options[:associated_with]}" if @options.key?(:associated_with)
@@ -149,6 +151,8 @@ module Audited
         "Expected #{model_class} to not have associated audits"
       end
 
+      alias_method :failure_message_when_negated, :negative_failure_message
+
       def description
         "has associated audits"
       end
@@ -164,7 +168,7 @@ module Audited
       end
 
       def association_exists?
-        (!reflection.nil?) &&
+        !reflection.nil? &&
           reflection.macro == :has_many &&
           reflection.options[:class_name] == Audited.audit_class.name
       end
