@@ -2,23 +2,15 @@ require 'rails/generators'
 require 'rails/generators/migration'
 require 'active_record'
 require 'rails/generators/active_record'
+require 'generators/audited/migration'
 
 module Audited
   module Generators
     class UpgradeGenerator < Rails::Generators::Base
       include Rails::Generators::Migration
+      extend Audited::Generators::Migration
 
       source_root File.expand_path("../templates", __FILE__)
-
-      # Implement the required interface for Rails::Generators::Migration.
-      def self.next_migration_number(dirname) #:nodoc:
-        next_migration_number = current_migration_number(dirname) + 1
-        if ActiveRecord::Base.timestamped_migrations
-          [Time.now.utc.strftime("%Y%m%d%H%M%S"), "%.14d" % next_migration_number].max
-        else
-          "%.3d" % next_migration_number
-        end
-      end
 
       def copy_templates
         migrations_to_be_applied do |m|
@@ -29,8 +21,8 @@ module Audited
       private
 
       def migrations_to_be_applied
-        Audited::Adapters::ActiveRecord::Audit.reset_column_information
-        columns = Audited::Adapters::ActiveRecord::Audit.columns.map(&:name)
+        Audited::Audit.reset_column_information
+        columns = Audited::Audit.columns.map(&:name)
 
         unless columns.include?( 'comment' )
           yield :add_comment_to_audits
@@ -42,6 +34,10 @@ module Audited
 
         unless columns.include?( 'remote_address' )
           yield :add_remote_address_to_audits
+        end
+
+        unless columns.include?( 'request_uuid' )
+          yield :add_request_uuid_to_audits
         end
 
         unless columns.include?( 'association_id' )
